requirements.txt
libclang==11.1.0
typing==3.7.4.3
flake8==3.9.0
pylint==2.6.2
mypy==0.812
